#11: Change project structure
