--- src/app/app.component.css
@@ -1,25 +1,0 @@
-.selected {
-  background-color: #CFD8DC !important;
-  color: white;
-}
-h1, h2 {
-  text-align: center;
-}
-table {
-  width: 340px;
-  margin: auto;
-}
-.col0 {
-  width: 40px;
-}
-.col1 {
-  width: 200px;
-  text-align: center;
-}
-.col2 {
-  width: 100px;
-  text-align: center;
-}
-.check {
-  margin: 4px;
-}
