+++ src/app/app.component.html
@@ -1,22 +1,1 @@
-<h1>{{title}}</h1>
-<h2>People</h2>
-<table class="table table-bordered table-hover">
-  <thead>
-  <tr>
-    <th>#</th>
-    <th class="col0"></th>
-    <th class="col1" (click)="setKey('to')">ID</th>
-    <th class="col2" (click)="setKey('point')">Джуджики</th>
-  </tr>
-  </thead>
-  <tbody>
-  <tr *ngFor="let person of people | sorting: key : counter; let i = index "
-      [class.selected]="person === selectedPerson"
-      (click)="onSelect(person, i)">
-    <td>{{i+1}}</td>
-    <td><input class="check" type="checkbox"/></td>
-    <td>{{person.to}}</td>
-    <td class="col2">{{person.point}}</td>
-  </tr>
-  </tbody>
-</table>
+<gamification-table></gamification-table>
