+++ src/app/app.module.ts
@@ -1,11 +1,12 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule }   from '@angular/forms';
-import { SortingPipe }   from './sorting.pipe';
+import { SortingPipe }   from './pipes/sorting.pipe';
 import { HttpModule }    from '@angular/http';
 
 import { AppComponent }  from './app.component';
-import {HttpService}     from "./http.service";
+import {TableComponent} from  './components/table/table.component';
+import {HttpService}     from "./service/http.service";
 
 @NgModule({
   imports: [
@@ -15,7 +16,8 @@ import {HttpService}     from "./http.service";
   ],
   declarations: [
     AppComponent,
-    SortingPipe
+    SortingPipe,
+    TableComponent
   ],
   bootstrap: [
     AppComponent
