+++ src/app/pipes/sorting.pipe.ts
