+++ src/index.html
@@ -5,8 +5,12 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <base href="/">
-    <link rel="stylesheet" href="bootstrap.min.css">
+    <!--<link rel="stylesheet" href="bootstrap.min.css">-->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
     <link rel="stylesheet" href="styles.css">
+
+    <script src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
 
     <!-- Polyfill(s) for older browsers -->
     <script src="../node_modules/core-js/client/shim.min.js"></script>
@@ -16,13 +20,13 @@
 
     <script src="systemjs.config.js"></script>
     <script>
-      System.import('main.js').catch(function(err){ console.error(err); });
+      System.import('app/main.js').catch(function(err){ console.error(err); });
     </script>
   </head>
 
   <body>
-    <my-app>Loading AppComponent content here ...</my-app>
-    <script src="jquery-3.1.1.js"></script>
-    <script src="bootstrap.min.js"></script>
+    <my-app>Loading gamification content ...</my-app>
+    <!--<script src="jquery-3.1.1.js"></script>
+    <script src="bootstrap.min.js"></script>-->
   </body>
 </html>
